refactor(ui): MainMenu UI migration + prefabs and scene changes
