+++ Assets/Prefabs/UI/Panels.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 24251d57a0520c3daaa4f662f8e94b6e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
