+++ Assets/Prefabs/UI/Panels/FreeModeSetup.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e0cdac6cac83b04289c96be9c4c1cc03
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
