+++ Assets/Prefabs/UI/Panels/MainMenu.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e3cdd20bccf2c986a825c9b270d57d4a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
