+++ Assets/Prefabs/UI/Panels/MissionResolution.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c994723a108e707bda852a7872afd0af
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
