+++ Assets/Prefabs/UI/Panels/NewCampaign.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7f6a5b91305611e398429dda53bc8db3
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
