+++ Assets/Prefabs/UI/Panels/Notifications.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8d41f44450562ea93b063914a0c8da3e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
