+++ Assets/Prefabs/UI/Panels/Pause.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b5452539250c145a285e976e93dc5938
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
