+++ Assets/Prefabs/UI/Panels/ProgressionGameOver.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5811ef63ab5b19495852ad2a5ca659e2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
