+++ Assets/Prefabs/UI/Panels/ProgressionSlots.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cf3e414adef6859f2af6ac1a62b52293
+guid: 8cb99342838eeed5b97dca237a0d1a96
 PrefabImporter:
   externalObjects: {}
   userData:
