+++ Assets/Prefabs/UI/Panels/Settings.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 911fc516ba15f6b26a44f930f7b7c6e2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
