+++ Assets/Prefabs/UI/Panels/ShipLibrary.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0b4e054bc2e5d297b882580298ff7562
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
