+++ Assets/Scripts/UI/Scenes/MainMenu/MainMenu.uxml
@@ -19,77 +19,5 @@
                        style="margin-bottom: 8px;"/>
             <ui:Button name="quit-button" text="Quit" class="ds-btn ds-btn--danger ds-btn--block"/>
         </ui:VisualElement>
-
-        <ui:VisualElement name="progression-slots-overlay" class="ds-backdrop" style="display: none;">
-            <ui:VisualElement class="ds-dialog" style="width: 520px; max-width: 520px;">
-                <ui:Label text="Progression" class="ds-dialog__title"/>
-                <ui:Label text="Choose a save slot" class="ds-body-2 ds-text-secondary"
-                          style="margin-bottom: 16px; -unity-text-align: middle-center;"/>
-                <ui:VisualElement class="progression-slots-list">
-                    <ui:VisualElement name="progression-slot-0-row" class="progression-slot-row">
-                        <ui:Button name="progression-slot-0-button" text="New game"
-                                   class="ds-btn ds-btn--secondary progression-slot-button"/>
-                        <ui:Button name="progression-slot-0-delete-button"
-                                   class="ds-btn ds-btn--icon ds-btn--icon-danger progression-slot-delete-button">
-                            <ui:VisualElement class="ds-icon ds-icon--trash"/>
-                        </ui:Button>
-                    </ui:VisualElement>
-                    <ui:VisualElement name="progression-slot-1-row" class="progression-slot-row">
-                        <ui:Button name="progression-slot-1-button" text="New game"
-                                   class="ds-btn ds-btn--secondary progression-slot-button"/>
-                        <ui:Button name="progression-slot-1-delete-button"
-                                   class="ds-btn ds-btn--icon ds-btn--icon-danger progression-slot-delete-button">
-                            <ui:VisualElement class="ds-icon ds-icon--trash"/>
-                        </ui:Button>
-                    </ui:VisualElement>
-                    <ui:VisualElement name="progression-slot-2-row" class="progression-slot-row">
-                        <ui:Button name="progression-slot-2-button" text="New game"
-                                   class="ds-btn ds-btn--secondary progression-slot-button"/>
-                        <ui:Button name="progression-slot-2-delete-button"
-                                   class="ds-btn ds-btn--icon ds-btn--icon-danger progression-slot-delete-button">
-                            <ui:VisualElement class="ds-icon ds-icon--trash"/>
-                        </ui:Button>
-                    </ui:VisualElement>
-                </ui:VisualElement>
-                <ui:VisualElement class="ds-dialog__actions">
-                    <ui:Button name="progression-slots-cancel-button" text="Back" class="ds-btn ds-btn--secondary"/>
-                </ui:VisualElement>
-            </ui:VisualElement>
-        </ui:VisualElement>
-
-        <ui:VisualElement name="progression-delete-confirm-overlay" class="ds-backdrop" style="display: none;">
-            <ui:VisualElement class="ds-dialog" style="width: 420px; max-width: 420px;">
-                <ui:Label text="Delete save?" class="ds-dialog__title"/>
-                <ui:Label text="This progression save will be permanently deleted." class="ds-body-2"
-                          style="margin-bottom: 16px;"/>
-                <ui:VisualElement class="ds-dialog__actions">
-                    <ui:Button name="progression-delete-cancel-button" text="Cancel" class="ds-btn ds-btn--secondary"/>
-                    <ui:Button name="progression-delete-confirm-button" text="Delete" class="ds-btn ds-btn--danger"/>
-                </ui:VisualElement>
-            </ui:VisualElement>
-        </ui:VisualElement>
-
-        <ui:VisualElement name="ship-select-overlay" class="ds-backdrop" style="display: none;">
-            <ui:VisualElement name="ship-select-dialog" class="ds-dialog" style="width: 520px; max-width: 520px;">
-                <ui:Label text="Free Mode" class="ds-dialog__title"/>
-                <ui:DropdownField name="ship-dropdown" label="Snapshot" class="ds-dropdown"
-                                  style="margin-bottom: 8px; min-width: 450px;"/>
-                <ui:Label text="Saved snapshots come from ShipFactory." class="ds-body-2 ds-text-primary"
-                          style="margin-bottom: 8px; opacity: 0.9;"/>
-                <ui:Label text="Asteroids" class="ds-body-2 ds-text-primary" style="margin-bottom: 2px;"/>
-                <ui:Slider name="asteroid-count-slider" low-value="0" high-value="30" show-input-field="true"
-                           class="ds-slider" style="margin-bottom: 8px; width: 100%;"/>
-                <ui:Label text="Enemy Ships" class="ds-body-2 ds-text-primary" style="margin-bottom: 2px;"/>
-                <ui:Slider name="enemy-count-slider" low-value="0" high-value="15" show-input-field="true"
-                           class="ds-slider" style="margin-bottom: 8px; width: 100%;"/>
-                <ui:Label text="Friendly Ships" class="ds-body-2 ds-text-primary" style="margin-bottom: 2px;"/>
-                <ui:Slider name="friendly-count-slider" low-value="0" high-value="15" show-input-field="true"
-                           class="ds-slider" style="margin-bottom: 32px; width: 100%;"/>
-                <ui:VisualElement class="ds-dialog__actions">
-                    <ui:Button name="ship-select-cancel-button" text="Cancel" class="ds-btn ds-btn--secondary"/>
-                    <ui:Button name="ship-select-launch-button" text="Launch" class="ds-btn ds-btn--primary"/>
-                </ui:VisualElement>
-            </ui:VisualElement>
-        </ui:VisualElement>
     </ui:VisualElement>
 </ui:UXML>
