+++ Assets/Scripts/UI/Scenes/MainMenu/Views/FreeMode.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a47a6f48216abfd278bdc24154cddad1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
