+++ Assets/Scripts/UI/Scenes/MainMenu/Views/FreeMode/FreeModeSetup.uxml
@@ -1,0 +1,30 @@
+<ui:UXML xmlns:ui="UnityEngine.UIElements" editor-extension-mode="False">
+    <Style src="project://database/Assets/DesignSystem/Resources/UI/Styles/DesignSystem/DesignSystem.uss"/>
+    <Style src="project://database/Assets/Scripts/UI/Scenes/MainMenu/MainMenu.uss"/>
+
+    <ui:VisualElement name="free-mode-setup-root" class="ds-root ds-root--transparent"
+                      style="position: absolute; left: 0; right: 0; top: 0; bottom: 0;">
+        <ui:VisualElement name="ship-select-overlay" class="ds-backdrop" style="display: flex;">
+            <ui:VisualElement name="ship-select-dialog" class="ds-dialog" style="width: 520px; max-width: 520px;">
+                <ui:Label text="Free Mode" class="ds-dialog__title"/>
+                <ui:DropdownField name="ship-dropdown" label="Snapshot" class="ds-dropdown"
+                                  style="margin-bottom: 8px; min-width: 450px;"/>
+                <ui:Label text="Saved snapshots come from ShipFactory." class="ds-body-2 ds-text-primary"
+                          style="margin-bottom: 8px; opacity: 0.9;"/>
+                <ui:Label text="Asteroids" class="ds-body-2 ds-text-primary" style="margin-bottom: 2px;"/>
+                <ui:Slider name="asteroid-count-slider" low-value="0" high-value="30" show-input-field="true"
+                           class="ds-slider" style="margin-bottom: 8px; width: 100%;"/>
+                <ui:Label text="Enemy Ships" class="ds-body-2 ds-text-primary" style="margin-bottom: 2px;"/>
+                <ui:Slider name="enemy-count-slider" low-value="0" high-value="15" show-input-field="true"
+                           class="ds-slider" style="margin-bottom: 8px; width: 100%;"/>
+                <ui:Label text="Friendly Ships" class="ds-body-2 ds-text-primary" style="margin-bottom: 2px;"/>
+                <ui:Slider name="friendly-count-slider" low-value="0" high-value="15" show-input-field="true"
+                           class="ds-slider" style="margin-bottom: 32px; width: 100%;"/>
+                <ui:VisualElement class="ds-dialog__actions">
+                    <ui:Button name="ship-select-cancel-button" text="Cancel" class="ds-btn ds-btn--secondary"/>
+                    <ui:Button name="ship-select-launch-button" text="Launch" class="ds-btn ds-btn--primary"/>
+                </ui:VisualElement>
+            </ui:VisualElement>
+        </ui:VisualElement>
+    </ui:VisualElement>
+</ui:UXML>
