+++ Assets/Scripts/UI/Scenes/MainMenu/Views/FreeMode/FreeModeSetup.uxml.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: d44e05c23ec1b6cda821feea6acb3ddd
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 13804, guid: 0000000000000000e000000000000000, type: 0}
