+++ Assets/Scripts/UI/Scenes/MainMenu/Views/FreeMode/FreeModeSetupController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: dd95a067bffd202f3bafda42c844bd46
