+++ Assets/Scripts/UI/Scenes/MainMenu/Views/Progression/ProgressionSlots.uxml.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: b7d3b30408fc224a595c51fef1304515
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 13804, guid: 0000000000000000e000000000000000, type: 0}
