+++ Assets/Scripts/UI/Scenes/MainMenu/Views/Progression/ProgressionSlotsController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: bac3d18c7e4b1fa7189ef8b3419523ac
